requirements.txt
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
flake8==7.0.0
mypy==1.10.0

# Trading and Market Data
ccxt>=4.0.0
pandas>=1.5.0
yfinance>=0.2.18

# Machine Learning
tensorflow>=2.13.0
keras>=2.13.1
transformers>=4.30.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Image Processing
matplotlib>=3.7.1
opencv-python>=4.8.0
Pillow>=10.0.0

# Technical Analysis
ta>=0.10.2

# Social Media and NLP
tweepy>=4.14.0
praw>=7.7.0
textblob>=0.17.1
nltk>=3.8.1

# Database - PostgreSQL
psycopg2-binary>=2.9.7
asyncpg>=0.28.0
SQLAlchemy>=2.0.0
alembic>=1.12.0

# Database - Redis
redis>=4.6.0
aioredis>=2.0.1

# Web Framework
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
cryptography>=41.0.0

# HTTP and API
httpx>=0.24.0
aiohttp>=3.8.5
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
schedule>=1.2.0
celery>=5.3.0
python-dateutil>=2.8.2

# Development
black>=23.0.0
isort>=5.12.0
pre-commit>=3.4.0

# Event-driven
pyee[asyncio]>=9.0.0

# Telegram Bot
aiogram>=3.0.0

# RAG/LLM (если используешь)
langchain>=0.1.0
llama-index>=0.9.0
openai>=1.0.0 